+++ apps/api/src/integration/integration.controller.ts
@@ -18,6 +18,7 @@ import { Authority, Integration } from '@prisma/client'
 import { RequiredApiKeyAuthorities } from '@/decorators/required-api-key-authorities.decorator'
 import { UpdateIntegration } from './dto/update.integration/update.integration'
 import { AuthenticatedUser } from '@/user/user.types'
+import { GetVercelEnvironments } from './dto/getVercelEnvironments/getVercelEnvironments'
 
 @Controller('integration')
 export class IntegrationController {
@@ -140,5 +141,18 @@ export class IntegrationController {
       user,
       integrationSlug
     )
+  }
+
+  @Put('vercel/environments')
+  @RequiredApiKeyAuthorities(
+    Authority.READ_PROJECT,
+    Authority.READ_ENVIRONMENT,
+    Authority.READ_INTEGRATION
+  )
+  async getVercelEnvironments(
+    @CurrentUser() user: AuthenticatedUser,
+    @Body() dto: GetVercelEnvironments
+  ) {
+    return await this.integrationService.getVercelEnvironments(user, dto)
   }
 }